chore: e2e testing rework
