+++ client/e2e/fixtures/fixtures.ts
@@ -1,6 +1,74 @@
-import { mergeTests, mergeExpects } from '@playwright/test';
-import { expect as axeExpect, test as axeTest } from './axe';
-import { test as configurationTest } from './setup';
+import {
+  mergeTests,
+  mergeExpects,
+  expect as baseExpect,
+  test as baseTest,
+} from '@playwright/test';
+import { ConfigurationPage } from '../pages/ConfigurationPage';
+import { ConfigurationsPage } from '../pages/ConfigurationsPage';
+import { AxeBuilder } from '@axe-core/playwright';
 
-export const expect = mergeExpects(axeExpect);
-export const test = mergeTests(axeTest, configurationTest);
+type Fixtures = {
+  configurationPage: ConfigurationPage;
+  configurationsPage: ConfigurationsPage;
+  makeAxeBuilder: () => AxeBuilder;
+};
+
+// See here: https://playwright.dev/docs/accessibility-testing#creating-a-fixture
+const extendedTest = baseTest.extend<Fixtures>({
+  makeAxeBuilder: async ({ page }, use) => {
+    const makeAxeBuilder = () =>
+      new AxeBuilder({ page }).withTags(['wcag21aa']);
+    await use(makeAxeBuilder);
+  },
+  configurationPage: async ({ page }, use) => {
+    await use(new ConfigurationPage(page));
+  },
+  configurationsPage: async ({ page }, use) => {
+    await use(new ConfigurationsPage(page));
+  },
+});
+
+const extendedExpect = baseExpect.extend({
+  async toHaveNoAxeViolations(makeAxeBuilder: () => AxeBuilder) {
+    const assertionName = 'toHaveNoAxeViolations';
+    const axe = makeAxeBuilder();
+    let pass: boolean;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    let results: any;
+    try {
+      results = await axe.analyze();
+    } catch {
+      pass = false;
+    }
+    const violations = results.violations ?? [];
+
+    pass = violations.length === 0;
+
+    const message = pass
+      ? () =>
+          this.utils.matcherHint(assertionName, undefined, undefined, {
+            isNot: this.isNot,
+          }) +
+          '\n\n' +
+          'No accessibility violations found'
+      : () =>
+          this.utils.matcherHint(assertionName, undefined, undefined, {
+            isNot: this.isNot,
+          }) +
+          '\n\n' +
+          `Found ${violations.length} accessibility violation(s):\n` +
+          JSON.stringify(violations, null, 2);
+
+    return {
+      message,
+      pass,
+      name: assertionName,
+      expected: 'no axe violations',
+      actual: `${violations.length} violation(s)`,
+    };
+  },
+});
+
+export const expect = mergeExpects(extendedExpect);
+export const test = mergeTests(extendedTest);
