+++ client/playwright.config.ts
@@ -9,7 +9,7 @@ export default defineConfig({
   },
   retries: process.env.CI ? 2 : 0,
   forbidOnly: !!process.env.CI,
-  workers: process.env.CI ? 1 : 2,
+  workers: 1,
   reporter: [['html', { open: 'never' }]],
   use: {
     trace: 'on-first-retry',
